+++ packages/medusa/src/api/store/products/route.ts
@@ -14,7 +14,11 @@ export const GET = async (
   req: RequestWithContext<HttpTypes.StoreProductListParams>,
   res: MedusaResponse<HttpTypes.StoreProductListResponse>
 ) => {
-  const filterableFields: HttpTypes.StoreProductListParams =
+  const filterableFields: HttpTypes.StoreProductListParams & {
+    // these are available through the Zod transformation
+    tags?: string | string[]
+    categories?: string | string[]
+  } =
     req.filterableFields
 
   if (FeatureFlag.isFeatureEnabled(IndexEngineFeatureFlag.key)) {