+++ www/utils/packages/docs-generator/src/classes/kinds/oas.ts
@@ -1777,7 +1777,7 @@ class OasKindGenerator extends FunctionKindGenerator {
         }
       case itemType.isIntersection():
         const allOfItems = this.typesHelper
-          .cleanUpTypes((itemType as ts.IntersectionType).types)
+          .cleanUpIntersectionTypes((itemType as ts.IntersectionType).types)
           .map((intersectionType) => {
             return this.typeToSchema({
               itemType: intersectionType,